requirements.txt
# Web Framework
streamlit==1.42.2
fastapi==0.115.11
uvicorn==0.34.0

# LangChain & LangGraph
langchain==0.3.19
langchain-openai==0.3.7
langchain-community==0.3.18
langgraph==0.3.2

# Azure OpenAI
openai==1.65.2

# Vector Database
faiss-cpu==1.10.0

# Utilities
python-dotenv==1.0.1
tiktoken==0.9.0
requests==2.32.3
pydantic==2.10.5
pydantic-settings==2.7.1

# Data Processing
beautifulsoup4==4.13.3
pandas==2.0.3

# Database
SQLAlchemy>=2.0.0

# Scheduling
APScheduler==3.10.4

# Progress Bar
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Observability (Optional)
langfuse==2.59.7

# Visualization
matplotlib==3.10.1
networkx==3.4.2
